backend/requirements.txt
# LinkOps Backend API - Python Dependencies
# Python 3.11+ required

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Async Support
aiosqlite==0.19.0
asyncio==3.4.3

# SSH Client
asyncssh==2.14.1

# Git Operations
GitPython==3.1.40

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0  # Must be <5.0 for passlib 1.7.4 compatibility
python-dotenv==1.0.0
pydantic[email]  # For email validation

# YAML Parsing
PyYAML==6.0.1

# Database
sqlalchemy==2.0.23

# HTTP Client (for testing)
httpx==0.25.2

# WebSocket Support
websockets==12.0

# Utilities
python-dateutil==2.8.2
pydantic==2.5.2
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1

# Development Tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Logging
structlog==23.2.0
python-json-logger==2.0.7
